==> scripts/In_process/4-Across tissues/4.0-Cluster refinement for Slingshot.R <==
# ============================================================================
# Name: Cluster refinement
# Author: elopez
# Date: May-15-2024
# Description: 
# TODO: 
# ============================================================================
library(Seurat)
library(dplyr)
library(ggplot2)
library(RColorBrewer)
# ============================================================================
# Section 1: CLUSTER REFINEMENT
# ============================================================================
# Loading seurat object with all tissues
S.Obj <- readRDS(file = "D:/R_ScriptsPaper/Objects/AcrossTissues_Wu.rds")

Idents(S.Obj)<-S.Obj$seurat_clusters
#REFINAR CLUSTERS NUMERICOS SEPARADOS
num_cl<-levels(unique(S.Obj$seurat_clusters))
high_list<-list()
#Plots highlighting each cluster
for (c in 1:length(num_cl)) {                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                          
  high_list[[c]]<-DimPlot(S.Obj, 
                          cells.highlight = colnames(S.Obj)[S.Obj$seurat_clusters==(c-1)])+
    NoLegend()+
    ggtitle(num_cl[c])+
    theme(axis.title = element_blank(),
          axis.text = element_blank(),
          plot.title = element_text(size = rel(0.5)))
}
#TISSUE
gridExtra::grid.arrange(grobs=high_list, ncol=5)

#Using cell selector
# Refinar C1 Parte superior de C2 y C8
A <- CellSelector(plot = DimPlot(subset(S.Obj, idents = c(2,8)), 
                                 group.by = "seurat_clusters",label = T)+NoLegend())
# Refinar C5. Parte inferior de C2
B <- CellSelector(plot = DimPlot(subset(S.Obj, idents = c(2)), 
                                 group.by = "seurat_clusters",label = T)+NoLegend())
# Refinar C9
C <- CellSelector(plot = DimPlot(subset(S.Obj, idents = 10), 
                                 group.by = "seurat_clusters",label = T)+NoLegend())
# Refinar C13
D <- CellSelector(plot = DimPlot(subset(S.Obj, idents = 10), 
                                 group.by = "seurat_clusters",label = T)+NoLegend())

#Elimination intersections
ABCD_venn<-venn::venn(list(A,B,C,D))


Cells_ID<-c(setdiff(A,B),B,C,D)

NewIdents<-c(rep("1",length(setdiff(A,B))),
             rep("5",length(B)),
             rep("9",length(C)),
             rep("13",length(D)))

S.Obj <- SetIdent(S.Obj, cells = Cells_ID, value = NewIdents)
S.Obj$seurat_clusters<-Idents(S.Obj)

# saveRDS(S.Obj, file = "D:/R_ScriptsPaper/Def_Objects/ClustRefin.rds")
# ============================================================================
# Section 2: DETERMINING MOST ABUNDANT PHENOTYPE PER CLUSTER
# ============================================================================
ClustRefin <-readRDS(file = "D:/R_ScriptsPaper/Def_Objects/ClustRefin.rds")
ClustRefin$seurat_clusters <- paste0("C",ClustRefin$seurat_clusters)

#ASSIGNING MOST ABUNDANT PHENOTYPE
all_cl<-table(ClustRefin$seurat_clusters)

#FOR SECOND ROUND
#List of cells with high ES
HIGH_LIST <- readRDS(file = "D:/R_ScriptsPaper/Def_Objects/TissueScores_HighCells_List.rds")

#HIGH_LIST$Unassigned <- setdiff(colnames(WuTAMs),unique(unlist(HIGH_LIST)))

#Transform list into named vector
my_vector <- unlist(HIGH_LIST)
names(my_vector) <- rep(names(HIGH_LIST), lengths(HIGH_LIST))

# Initialize placeholders
substitutions <- vector()
overlap.map <- list()
# Define threshold to name cluster
threshold <- 33.33
for (ii in seq_along(all_cl)) {
  # Get the names that are in the specified cluster
  matched_names <- names(my_vector)[
    my_vector %in% colnames(ClustRefin)[ClustRefin$seurat_clusters == names(all_cl[ii])]]
  # Calculate % that each phenotype occupies in each cluster
  percent <- table(matched_names) / all_cl[ii] * 100
  sorted_values <- percent[order(percent, decreasing = TRUE)]
  sorted_names <- names(sorted_values)
  # The Highest phenotype should be at least 1/3 of the cluster
  if (length(sorted_values) > 0 && sorted_values[1] >= threshold) {
    substitutions[ii] <- sorted_names[1]
    # If the second higher is also above 33.33%, create a list
    # with the respective cluster and the name of the second highest phenotype as element
    if (length(sorted_values) > 1 && sorted_values[2] >= threshold) {
      overlap.map[[names(all_cl[ii])]] <- sorted_names[2]
    }
  } else {
    # If no phenotype achieve parameters. Label as unassigned
    substitutions[ii] <- "Unassigned"
    # Rescue in unassigned
    overlap.map[[names(all_cl[ii])]] <- sorted_names[1]
    # Overlap with the highest of 2 highest if difference is less than 5%
    if (sorted_values[1]-sorted_values[2] <=5 ){
      overlap.map[[names(all_cl[ii])]] <-c(sorted_names[1],sorted_names[2])
    }
  }
}
names(substitutions) <- names(all_cl)

# Replacing cluster number by the most abundant phenotype in the cluster
ClustRefin$MainPheno <- dplyr::recode(ClustRefin$seurat_clusters, !!!substitutions)


# RESCUE PHENOTYPES
# Rescue phenotypes (as names) and set respective overlaping phenos
for (ii in seq_along(overlap.map)) {
  # Cells within the clusters
  cl_logic<-ClustRefin$seurat_clusters%in%names(overlap.map[ii])
  # Cells with high of the respective phenotype
  High_logic<-colnames(ClustRefin)%in%as.character(unlist(HIGH_LIST[overlap.map[[ii]]]))
  # Cells with both
  logical_vect <-cl_logic&High_logic
  # Substituing
  ClustRefin$MainPheno<-ifelse(logical_vect,
                           overlap.map[[ii]],
                           as.character(ClustRefin$MainPheno))
}

#Distinct <- DiscretePalette(length(unique(ClustRefin$seurat_clusters)),shuffle = T)
Distinct <- c("#FFE100", #APOE
              "#0075DC", #CXCL9
              "#2BCE48", #FCN1
              "#FF0010", #FOLR2
              "#993F00", #FOSB
              "#FF5005", #IL1B
              "#F0A0FF", #ISG15
              "darkgrey" #Unassigned
              )

#Replacing CD16 and Cycling TAMs
#ClustRefin$MainPheno<-ifelse(ClustRefin$celltypes%in%c("CD16 monocytes","Cycling TAMs"),
 #                             as.character(ClustRefin$celltypes),
  #                            as.character(ClustRefin$MainPheno))

#Clusters que abarca cada fenotipo
pA <- DimPlot(ClustRefin, group.by = "seurat_clusters",label = T,pt.size =0.5, 
            cols = DiscretePalette(length(unique(ClustRefin$seurat_clusters)),shuffle = T))+ 
  NoLegend()

pB <- DimPlot(ClustRefin, group.by = "MainPheno", pt.size =0.5,
            cols = Distinct)
pA|pB
# Save object
# saveRDS(ClustRefin,file = "D:/R_ScriptsPaper/Def_Objects/MainPhenotype.rds")
